refactor: Extract view model and rename view
