+++ SetGame/Programming_Assignment_3/SetGame/SetGame/SetGameApp.swift
@@ -14,7 +14,7 @@ struct SetGameApp: App {
     
     var body: some Scene {
         WindowGroup {
-            ContentView(viewModel: game)
+            SetGameView(viewModel: game)
         }
     }
 }
