+++ SetGame/Programming_Assignment_3/SetGame/SetGame/SetGameView.swift
@@ -1,5 +1,5 @@
 //
-//  ContentView.swift
+//  SetGameView.swift
 //  SetGame
 //
 //  Created by Jill Chang on 2024/6/24.
@@ -7,7 +7,7 @@
 
 import SwiftUI
 
-struct ContentView: View {
+struct SetGameView: View {
     @ObservedObject
     var viewModel: SetGameViewModel
     
@@ -46,36 +46,5 @@ struct ContentView: View {
 }
 
 #Preview {
-    ContentView(viewModel: SetGameViewModel())
-}
-
-class SetGameViewModel: ObservableObject {
-    @Published
-    private var model = SetGame()
-    var isGameOver = false
-    
-    var cards: [SetGame.Card] {
-        model.cards
-    }
-    
-    var score: Int {
-        model.score
-    }
-    
-    var deckIsEmpty: Bool {
-        model.deck.isEmpty
-    }
-    
-    func selectCard(_ card: SetGame.Card) {
-        // TODO: model.selectCard(card)
-        // TODO: check game is over?
-    }
-    
-    func dealThreeMoreCards() {
-        // TODO: model.dealThreeMoreCards()
-    }
-    
-    func newGame() {
-        model = SetGame()
-    }
+    SetGameView(viewModel: SetGameViewModel())
 }
