+++ SetGame/Programming_Assignment_3/SetGame/SetGame/SetGameViewModel.swift
@@ -1,0 +1,39 @@
+//
+//  SetGameViewModel.swift
+//  SetGame
+//
+//  Created by Jill Chang on 2024/6/26.
+//
+
+import Foundation
+
+class SetGameViewModel: ObservableObject {
+    @Published
+    private var model = SetGame()
+    var isGameOver = false
+    
+    var cards: [SetGame.Card] {
+        model.cards
+    }
+    
+    var score: Int {
+        model.score
+    }
+    
+    var deckIsEmpty: Bool {
+        model.deck.isEmpty
+    }
+    
+    func selectCard(_ card: SetGame.Card) {
+        // TODO: model.selectCard(card)
+        // TODO: check game is over?
+    }
+    
+    func dealThreeMoreCards() {
+        // TODO: model.dealThreeMoreCards()
+    }
+    
+    func newGame() {
+        model = SetGame()
+    }
+}
